% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AIC.streg}
\alias{AIC.streg}
\alias{AIC.summary.streg}
\alias{BIC.streg}
\alias{BIC.summary.streg}
\title{Akaike's Information Criterion}
\usage{
\method{AIC}{streg}(object, ..., k = 2)

\method{AIC}{summary.streg}(object, ...)

\method{BIC}{streg}(object, ...)

\method{BIC}{summary.streg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{streg} or \code{summary.streg}.}

\item{...}{Not used.}

\item{k}{The penalty per parameter to be used; defaults to $k = 2$, corresponding to the classical AIC.}
}
\description{
Extract Akaike's Information Criterion (AIC) or Bayesian Information Criterion (BIC) from fitted models of class \code{streg}.
}
