% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-methods.R
\name{coef.streg}
\alias{coef.streg}
\title{Extract Model Coefficients}
\usage{
\method{coef}{streg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{streg}.}

\item{...}{Not used.}
}
\description{
Extract model coefficients from a \code{streg} fit.
}
\seealso{
\code{\link[stats:coef]{stats::coef()}}

\code{\link[stats:coef]{stats::coefficients()}}
}
