% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hip}
\alias{hip}
\title{Hip Fracture Study}
\format{
A data frame with 206 rows and 8 variables:
\itemize{
\item \code{id} Patient ID;
\item \code{time0} Begin of time interval;
\item \code{time1} End of time interval;
\item \code{fracture} Fracture event;
\item \code{protect} Wears device;
\item \code{age} Age at enrollment;
\item \code{calcium} Blood calcium level;
\item \code{male} 1 if male.
}
}
\usage{
hip
}
\description{
Hip fracture study data, imported from Stata 16.
}
\examples{
data("hip", package = "streg")
}
\references{
http://www.stata-press.com/data/r16/hip3.dta
}
\keyword{datasets}
