% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.streg}
\alias{is.streg}
\alias{is.summary.streg}
\title{Parametric Survival Models}
\usage{
is.streg(x)

is.summary.streg(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Reports whether \code{x} is an object of class \code{streg} or \code{summary.streg}.
}
