% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kva}
\alias{kva}
\title{Generator Experiment}
\format{
A data frame with 12 rows and 4 variables:
\itemize{
\item \code{failtime} Time until failure (in hours);
\item \code{event} Event indicator variable;
\item \code{load} Overload (in KVA);
\item \code{bearings} Binary variable for \emph{Has new bearings?}.
}
}
\usage{
kva
}
\description{
Generator Experiment
}
\examples{
data("kva", package = "streg")
}
\keyword{datasets}
