% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logLik.streg}
\alias{logLik.streg}
\alias{logLik.summary.streg}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{streg}(object, ...)

\method{logLik}{summary.streg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{streg} or \code{summary.streg}.}

\item{...}{Not used.}
}
\description{
Extract log-likelihood of a \code{streg} model.
}
