% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streg-S3-methods.R
\name{nobs.streg}
\alias{nobs.streg}
\title{Extract the Number of Observations from a Fit.}
\usage{
\method{nobs}{streg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{streg}.}

\item{...}{Not used.}
}
\description{
Extract the number of 'observations' from an \code{streg} model fit.
}
\seealso{
\code{\link[stats:nobs]{stats::nobs()}}
}
