% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.streg}
\alias{plot.streg}
\title{Plot Survivor, Hazard, Cumulative Hazard, or Cumulative Incidence Function}
\usage{
\method{plot}{streg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{streg}.}

\item{...}{Not used.}
}
\description{
Plot Survivor, Hazard, Cumulative Hazard, or Cumulative Incidence Function
}
