% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streg-S3-methods.R
\name{print.streg}
\alias{print.streg}
\title{Print \code{streg} Fits}
\usage{
\method{print}{streg}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{streg}.}

\item{digits}{The number of significant digits to use when printing.}

\item{...}{Not used.}
}
\description{
Print the coefficients from a \code{streg} fit.
}
