% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streg-package.R
\docType{package}
\name{streg-package}
\alias{streg-package}
\alias{streg}
\title{Parametric Survival Analysis}
\description{
Survival analysis with parametric models.

Survival analysis parametric proportional hazards survival models, including
  the exponential, Weibull, and Gompertz model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ellessenne.github.io/streg}
  \item Report bugs at \url{https://github.com/ellessenne/streg/issues}
}

}
\author{
Alessandro Gasparini (alessandro.gasparini@ki.se)

\strong{Maintainer}: Alessandro Gasparini \email{alessandro.gasparini@ki.se} (\href{https://orcid.org/0000-0002-8319-7624}{ORCID})

}
\keyword{internal}
