% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.streg.R
\name{summary.streg}
\alias{summary.streg}
\alias{print.summary.streg}
\title{Summarizing \code{streg} Fits}
\usage{
\method{summary}{streg}(object, conf.int = 0.95, ...)

\method{print}{summary.streg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{An object of class \code{streg}}

\item{conf.int}{Significancy level for confidence intervals. Defaults to 0.95.}

\item{...}{Not used.}

\item{x}{An object of class \code{summary.streg}}

\item{digits}{The number of significant digits to use when printing.}
}
\description{
These functions are all methods for class \code{streg} or \code{summary.streg} objects.
}
