% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov.streg}
\alias{vcov.streg}
\alias{vcov.summary.streg}
\title{Calculate Variance-Covariance Matrix for a \code{streg} Model Object}
\usage{
\method{vcov}{streg}(object, ...)

\method{vcov}{summary.streg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{streg} or \code{summary.streg}.}

\item{...}{Not used.}
}
\description{
Returns the variance-covariance matrix of all estimated parameters of a fitted \code{streg} model.
}
