% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.streg}
\alias{coef.streg}
\alias{coef.summary.streg}
\title{Extract Model Coefficients}
\usage{
\method{coef}{streg}(object, ...)

\method{coef}{summary.streg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{streg} or \code{summary.streg}.}

\item{...}{Not used.}
}
\description{
\code{coef} extracts model coefficients from a \code{streg} model fit. \code{coefficients} is an alias for it.
}
