% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streg-S3-methods.R
\name{logLik.streg}
\alias{logLik.streg}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{streg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{streg}.}

\item{...}{Not used.}
}
\description{
Extract log-likelihood from a fitted \code{streg} model.
}
\seealso{
\code{\link[stats:logLik]{stats::logLik()}}
}
