% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streg-S3-methods.R
\name{vcov.streg}
\alias{vcov.streg}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{streg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{streg}.}

\item{...}{Not used.}
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted \code{streg} model.
}
